requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2.16
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
isort==5.10.1
mccabe==0.6.1
Pillow==8.3.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.9
